Support long numbers instead of integers
Even long is not enough in the real scientific world. BigInteger would be a next step,

+++ src/main/java/ro/gal/perfectnumber/rest/PerfectNumberController.java
@@ -19,7 +19,7 @@ public class PerfectNumberController {
     }
 
     @RequestMapping("/check")
-    public @ResponseBody boolean checkNumber(@RequestParam("number") int number) {
+    public @ResponseBody boolean checkNumber(@RequestParam("number") long number) {
         return service.isPerfectNumber(number);
     }
 }
